Got almost everything except marking finishedp3 todos.py
