+++ todos.py
@@ -9,12 +9,17 @@ def help_screen():
     print("""\nSelect an option.
           Type \"write\" to create a task.
           Type \"help\" to view this help screen.
-          Type \"quit\" to exit.\n
-          Type \"destroy\" to kill all tasks.""")
+          Type \"quit\" to exit.
+          Type \"destroy\" to kill all tasks.\n""")
 
-def list_all():
-    pass
-
+print("""
+*******************
+*   GITTER DONE   *
+*                 *
+* repl based to-  *
+* do list.        *
+*******************
+""")
 help_screen()
 manage_option=""
 options = ['write', 'help']
@@ -23,21 +28,15 @@ while True:
     manage_option= input("Type an action or \"help\". >")
 
     if manage_option.lower() == "write":
-        # Let's not learn to run before we learn to walk. We can always clear the file.
-        # eval is a thing...?
-        now_now= datetime.datetime.now()
-        time_stamp= f"{now_now.month}/{now_now.day}/{now_now.year}"
-        print("\n==Write New Task==\n")
-        print(time_stamp)
-        current_task_name = input("Task Name: ")
-        current_task_desc = input("Task Description: ")
-        current_task = manager.Manager(current_task_name, current_task_desc, time_stamp, False)
-        current_task.write_task()
+        # Put this in the Manager Module!
+        manager.create_task()
     elif manage_option.lower() == "help":
         help_screen()
     elif manage_option.lower() == "list":
-        pass
+        manager.list_all()
     elif manage_option.lower() == "destroy":
+        manager.destroy_tasks()
+    elif manager_option.lower() == "mark":
         pass
     elif manage_option.lower() == "quit":
         break
