--- todos.txt
@@ -1,3 +1,0 @@
-Buy a Dream Car, Buy a Chevvy Camaro even though I'm epileptic, 2/7/2019, False
-Go Outside, Go outside, 2/7/2019, False
-sfdgh, gfdh, 2/7/2019, False
